Rename files to more intuitive ones, easing newcomers adoption (like @inodb)
