+++ facs/Makefile
@@ -14,8 +14,8 @@ endif
 .PHONY: tests clean valgrind
 .SUFFIXES:.c .o
 PROG=facs
-LOBJS= big_query.o bloom.o file_dir.o good_build.o lookup8.o suggestions.o tool.o simple_check_1_ge.o simple_remove.o prob.o info.o
-AOBJS= big_query.o bloom.o file_dir.o good_build.o lookup8.o suggestions.o tool.o simple_check_1_ge.o simple_remove.o prob.o info.o
+LOBJS= query.o bloom.o file_dir.o build.o lookup8.o suggestions.o tool.o check.o remove.o prob.o info.o
+AOBJS= query.o bloom.o file_dir.o build.o lookup8.o suggestions.o tool.o check.o remove.o prob.o info.o
 
 all:$(PROG)
 
@@ -31,7 +31,7 @@ mpi:
 	#mpicc -c mpi_decon.c -O3 -D_FILE_OFFSET_BITS=64 -D_LARGE_FILE 
 	#mpicc -o mpi_decon mpi_decon.o bloom.o suggestions.o lookup3.o  -lm ${CFLAGS}
 	mpicc -c mpi_bloom.c ${CFLAGS}
-	mpicc -o facs_mpi tool.o big_query.o simple_check_1_ge.o simple_remove.o mpi_bloom.o bloom.o suggestions.o lookup8.o file_dir.o prob.o ${LIBS} ${CFLAGS}
+	mpicc -o facs_mpi tool.o query.o check.o remove.o mpi_bloom.o bloom.o suggestions.o lookup8.o file_dir.o prob.o ${LIBS} ${CFLAGS}
 
 python:
 	rm -rf build/ ${PROG}.so && python setup.py build_ext --inplace && python setup.py develop
@@ -51,5 +51,5 @@ ${PROG}:lib${PROG}.a $(AOBJS) main.o
 lib${PROG}.a:$(LOBJS)
 		$(AR) -csru $@ $(LOBJS)
 
-main.o: big_query.h bloom.h build.h check.h file_dir.h hashes.h tool.h remove.h info.h
+main.o: query.h bloom.h build.h check.h file_dir.h hashes.h tool.h remove.h info.h
