+++ facs/check.c
